+++ dependency-reduced-pom.xml
