+++ src/main/java/me/chunklock/ChunklockPlugin.java
@@ -8,10 +8,24 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.plugin.java.JavaPlugin;
-import me.chunklock.teams.EnhancedTeamManager;
-import me.chunklock.teams.BasicTeamCommandHandler;
-import me.chunklock.migration.DataMigrator;
+import me.chunklock.managers.EnhancedTeamManager;
+import me.chunklock.commands.BasicTeamCommandHandler;
+import me.chunklock.util.DataMigrator;
 import me.chunklock.commands.ChunklockCommandExecutor;
+import me.chunklock.commands.ChunklockCommand;
+import me.chunklock.managers.BiomeUnlockRegistry;
+import me.chunklock.managers.ChunkBorderManager;
+import me.chunklock.managers.ChunkEvaluator;
+import me.chunklock.managers.ChunkLockManager;
+import me.chunklock.managers.ChunkValueRegistry;
+import me.chunklock.managers.HologramManager;
+import me.chunklock.managers.PlayerDataManager;
+import me.chunklock.managers.PlayerProgressTracker;
+import me.chunklock.managers.TeamManager;
+import me.chunklock.managers.TickTask;
+import me.chunklock.listeners.BlockProtectionListener;
+import me.chunklock.listeners.PlayerListener;
+import me.chunklock.ui.UnlockGui;
 import java.util.logging.Level;
 
 import net.kyori.adventure.text.Component;
