+++ src/main/java/me/chunklock/managers/EnhancedTeamManager.java
@@ -1,4 +1,4 @@
-package me.chunklock.teams;
+package me.chunklock.managers;
 
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.ConfigurationSection;
@@ -6,6 +6,7 @@ import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
+import me.chunklock.models.Team;
 
 import java.io.File;
 import java.io.IOException;
