+++ src/main/java/me/chunklock/managers/PlayerDataManager.java
@@ -1,4 +1,4 @@
-package me.chunklock;
+package me.chunklock.managers;
 
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
