+++ src/main/java/me/chunklock/models/ChunkData.java
@@ -1,4 +1,4 @@
-package me.chunklock;
+package me.chunklock.models;
 
 import java.util.UUID;
