+++ src/main/java/me/chunklock/models/Difficulty.java
@@ -1,4 +1,4 @@
-package me.chunklock;
+package me.chunklock.models;
 
 public enum Difficulty {
     EASY,
