+++ src/main/java/me/chunklock/models/TeamRole.java
@@ -1,4 +1,4 @@
-package me.chunklock.teams;
+package me.chunklock.models;
 
 public enum TeamRole {
     OWNER("Owner"),
