+++ src/main/java/me/chunklock/util/DataMigrator.java
@@ -1,4 +1,4 @@
-package me.chunklock.migration;
+package me.chunklock.util;
 
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
